requirements.txt
Flask
Flask-MySQLdb
pandas
openpyxl
Pillow
Werkzeug
python-dotenv
flask-limiter
flask-wtf
rq
redis